refactor(common): Consolidate context menu components
