+++ packages/telestion-client-common/src/components/context-menu/index.ts
@@ -1,1 +1,2 @@
-export * from './context-menu';
+export * from './menu';
+export * from './wrappers';
