+++ packages/telestion-client-common/src/components/context-menu/menu/index.ts
@@ -1,0 +1,1 @@
+export * from './context-menu';
