+++ packages/telestion-client-common/src/components/context-menu/wrappers/index.ts
@@ -1,0 +1,2 @@
+export * from './context-menu-provider';
+export * from './context-menu-wrapper';
